requirements.txt
# Core ML/DL libraries
tensorflow==2.15.0
numpy==1.24.3
matplotlib==3.8.0
pillow==10.1.0

# Streamlit for web interface
streamlit==1.28.1

# Utilities
scikit-learn==1.3.2
